Subject: 2BSD patches... #448 From: Johnny Billquist <b...@softjar.se> Date: Sun, 17 Jun 2012 18:25:24 +0200
Hi. Here is a set of patches to 2BSD, which fixes a number of problems.
Terribly sorry that I don't present it in the same nice format that
Steven M. Schultz does, but I'll try an explain what this is briefly.

I've named this patch set path #448, as the last known patches to me are

Fixes in here:

====

1. Use of non-DEC MSCP controllers improved. Some parts of 2BSD have
been updated to work with (for example) CMD controllers, but not all
parts were. This set of patches makes it possible to boot and run with
the CDU-720, for example, which did not work before.

2. Boot program now automatically boots unless manual intervention on
console. This looks pretty similar to NetBSD on VAX for example, where a
countdown is presented at boot time, and the system continues with an
automatic boot unless aborted. Previously, 2BSD would not autoboot from
cold start because the reboot-flag was not present at power up.

3. Console terminal made 8-bit clean. On a real PDP-11, the boot
monitors are 8-bit clean. However, 2BSD previously ran with 7E on the
console, and there was no way to avoid this for system output. This
patch makes it all 8-bit clean.

4. The libc resolver code used /etc/hosts if no resolved was available,
but if one was, it never used the /etc/hosts. This created a peculiar
effect, especially at bootup, since the resolver couldn't be contacted
before the network was up, but /etc/hosts were not used, since a correct
/etc/resolv.conf existed. The order is not possible to select. It will
first try using the resolver, but if that fails, it now falls back to
trying /etc/hosts

5. At system build time, the newvers.sh tries to figure out various bits
and pieces to put into the built file to tell when the kernel was built,
where and by who. This parsing could fail in various ways because of how
the date command works with time zones. Fixed by changing how it figures
out the information and pass it around.

6. The mandoc macros had a Y2K bug, or rather a 2010 bug, in that the
Y2K bug fix actually only fixed years 2000-2009, and it broke again in
2010. This patch does a proper fix to the Y2K problem. Also fixed a
spelling error.

====

As usual, the code might not be pretty, but I've atleast been running it
myself on several machines for close to two years now, and believe these
are all workable, and important patches.
Download to your machine.
At the root of the file system run:
$ patch -p0 < patchfile

after this, rebuild the kernel and the boot image. Install the new
kernel, the new boot, and then rebuild all of userland.

If you have any questions, feel free to send me an email.

This patch set will bring your system up to patch version 448.

     Johnny

--
Johnny Billquist                  || "I'm on a bus
                                   ||  on a psychedelic trip
email: b...@softjar.se             ||  Reading murder books
pdp is alive!                     ||  tryin' to stay hip" - B. Idol